moved loadData.ts to api/
