+++ app/api/auth/[...nextauth]/route.ts
@@ -8,7 +8,7 @@ import useCurrentUser from "@/hooks/useCurrentUser";
 
 // Update this when switching databases if needed
 const supabaseUrl = 'https://kalbwmivszjzlnepcebm.supabase.co';
-const supabaseKey = process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY;
+const supabaseKey = process.env.SUPABASE_ANON_KEY;
 const supabase = createClient(supabaseUrl, supabaseKey!);
