+++ app/api/data/uploadSchoolData.ts
@@ -24,7 +24,7 @@ import { createClient } from "@supabase/supabase-js";
 
 
 const supabaseUrl = 'https://kalbwmivszjzlnepcebm.supabase.co';
-const supabaseKey = process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY;
+const supabaseKey = process.env.SUPABASE_ANON_KEY;
 const supabase = createClient(supabaseUrl, supabaseKey!);
 
 function ValidateColumns() {
