+++ next.config.js
@@ -1,8 +1,8 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   env: {
-    NEXT_PUBLIC_SUPABASE_ANON_KEY: process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY,
-    NEXT_PUBLIC_RESEND_API_KEY: process.env.NEXT_PUBLIC_RESEND_API_KEY
+    SUPABASE_ANON_KEY: process.env.SUPABASE_ANON_KEY,
+    RESEND_API_KEY: process.env.RESEND_API_KEY
   },
 };
